# api/openapi.yaml
openapi: 3.1.0
info:
  title: WorldBisect API
  version: 1.0.0
  description: >-
    Authenticated local API for captures, comparisons, analyses, jobs, reports,
    metrics, and audit verification. Remote command execution is disabled unless
    explicitly enabled in the daemon configuration.
servers:
  - url: http://127.0.0.1:8787
security:
  - bearerAuth: []
paths:
  /api/v1/health:
    get:
      operationId: health
      security: []
      responses:
        "200":
          description: Service health
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /api/v1/version:
    get:
      operationId: version
      responses:
        "200":
          description: Version information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Version"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/captures:
    get:
      operationId: listCaptures
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Capture summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Capture"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createCapture
      description: >-
        Queues a capture job. Requires the capture:write scope, remote execution
        enabled, an allowed canonical executable, and an allowed working directory.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureRequest"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/captures/{id}:
    get:
      operationId: getCapture
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Capture details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capture"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/analyses:
    get:
      operationId: listAnalyses
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Analysis summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAnalysis
      description: Queues a bounded comparison and causal analysis.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisRequest"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/analyses/{id}:
    get:
      operationId: getAnalysis
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Analysis details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/jobs:
    get:
      operationId: listJobs
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Job summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/jobs/{id}:
    get:
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/audit/verify:
    post:
      operationId: verifyAudit
      responses:
        "200":
          description: Audit-chain verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditVerification"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/metrics:
    get:
      operationId: metrics
      description: Prometheus text exposition.
      responses:
        "200":
          description: Metrics
          content:
            text/plain:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /:
    get:
      operationId: dashboard
      description: Embedded authenticated dashboard.
      responses:
        "200":
          description: Dashboard HTML
          content:
            text/html:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._-]+$"
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    Forbidden:
      description: Authenticated principal is not permitted to perform the operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    Conflict:
      description: Idempotency or state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    TooManyRequests:
      description: Quota exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
  schemas:
    APIError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    Version:
      type: object
      required: [version, commit, build_time]
      properties:
        version:
          type: string
        commit:
          type: string
        build_time:
          type: string
    CommandSpec:
      type: object
      required: [command, working_directory]
      properties:
        command:
          type: array
          minItems: 1
          items:
            type: string
        working_directory:
          type: string
        environment:
          type: object
          additionalProperties:
            type: string
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 3600
    Oracle:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [exit, timeout, stdout_regex, stderr_regex, file_digest]
        expected_exit_code:
          type: integer
        pattern:
          type: string
        file:
          type: string
        digest:
          type: string
    CaptureRequest:
      type: object
      required: [command, oracle]
      properties:
        command:
          $ref: "#/components/schemas/CommandSpec"
        oracle:
          $ref: "#/components/schemas/Oracle"
        label:
          type: string
          maxLength: 256
    AnalysisRequest:
      type: object
      required: [good_capture_id, bad_capture_id]
      properties:
        good_capture_id:
          type: string
        bad_capture_id:
          type: string
        repetitions:
          type: integer
          minimum: 1
          maximum: 100
          default: 3
    Capture:
      type: object
      additionalProperties: true
    Analysis:
      type: object
      additionalProperties: true
    Job:
      type: object
      additionalProperties: true
    AuditVerification:
      type: object
      required: [valid, entries]
      properties:
        valid:
          type: boolean
        entries:
          type: integer
        error:
          type: string